FIFO for asynchronous command response w/ gnugo
All commands are pushed to the back of a list as they are written to
gnugo. When data is read, it is appended to the first list element until
the complete response has been read, denoted by "\n\n". Once it is
complete it is popped off the list and distributed for processing. It is
slightly complicated but worth it since the GUI won't freeze while
waiting for gnugo to get reply.

Fixed inverted Y axis.

Added very crude response parser which knows the commands "play" and
"genmove" and places the moves on the board. Hacked it so that typing
"genmove black" in input box will cause it to autoplay.

The input box commands go straight to gnugo. Its responses show up in
the history text area. GoGoGo does nothing with commands until gnugo
responds.

Some kludges: Start gnugo by clicking the resign button. The Pass button
exits the gnugo process. Hint button drops a stone on the board.