White 1st move if handicap>1. Start file I/O.
This may also contain a temp fix to prevent crashing when final score is
tallied. For some reason it takes a very long time--window grays out.
Then the scores end up being accepted as hints. Hints expects
vertex-score pairs but the scores are 3 floats so it fails on a bounds
assertion. Got around it for now by checking for 2nd half of pair.

White is supposed to get the first move if there are > 1 handicap
stones.

Have it opening open and save file dialogs and looked up the gnugo
commands for loading and saving files. Shouldn't be a big deal to
finish.