Improved Hint marker text
Tried rotating the marker but for some reason that makes the rectangle
disappear. Ended up using random color and pen width for font when
marker is to the left(?) of another marker.

Also couldn't get qt's collision functions to work as I expected despite
trying all of the modes. So using contains() with a offset on x-axis.