Stone sound, stars, border
If there is a file called stone.wav in the same directory as the
executable, it will be played each time a stone is placed. I haven't
found a stone.wav that I'm certain is GPL so I don't want to commit it
to Git.

Added stars where handicap stones go.

Added thick border around board.